// qto-ui/src/app/models/service-cyber-view.model.ts
import { AbstractBaseModel } from "./abstract-base-model";

export class ServiceCyberView extends AbstractBaseModel {
  locationId: number;
  orderId: number;
  clientLocationId: number;
  parentCompanyName: string;
  parentCompanyClientId: string;
  companyName: string;
  endCustomerClientId: string;
  companyId: number;
  type: string;
  address: string;
  address1: string;
  address2: string;
  city: string;
  stateProvince: string;
  postalCode: string;
  status: string;
  followUpDate: Date | null;
  provisioner: string;
  i90ProjectManager: string;
  progressPercentage: number;
  provider: string;
  statusAge: number;
  mrc: number;
  nrc: number;
  greatestMilestoneName: string;
  greatestMilestoneDate: Date;
  latestNote: string;
  equipmentCount: number;
  equipmentTypes: string;
  customerRequestedInstall: Date;
  created: Date;
  techDataGatheringFormSent: Date;
  techDataGatheringMeetingScheduled: Date;
  techDataGatheringMeetingCompleted: Date;
  emailUsmAnywhereTemplateRequirementsSent: Date;
  inventoryAssignmentVerified: Date;
  newUsmAnywhereServerBuildStarted: Date;
  implementationQaCompleted: Date;
  verifyAssetsInSiemDb: Date;
  verifyLoggingDataSource: Date;
  scheduleVulnerabilityScans: Date;
  bulkAlarmTuningPhase1: Date;
  siemEventFiltering: Date;
  filtersBuiltForReports: Date;
  defaultAlarmRuleAdditions: Date;
  customAlarmRuleAdditions: Date;
  forwardAlarmsToUsmCentral: Date;
  forwardAlarmsToD3SocLive: Date;
  hostListProvidedByClient: Date;
  halcyonPackageGivenToClient: Date;
  halcyonDeployedToHosts: Date;
  devopsNotifiedOfHalcyonAddition: Date;
  halcyonApiTokenAddedToD3: Date;
  d3ConnectionVerified: Date;
  endLearningMode: Date;
  testEmailSentToClient: Date;
  deployConsultingTenant: Date;
  deployVulnerabilityScans: Date;
  providedCustomerWithReport: Date;
  discussFutureCyrismaManagement: Date;
  reviewExistingCaAndMfaPolicies: Date;
  signInPoliciesEnabled: Date;
  conditionalAccessPolicyVerification: Date;
  geographicRestrictionsEnabled: Date;
  deviceComplianceEnabled: Date;
  passwordResetEnabledForSelfService: Date;
  breakGlassAccountConfigured: Date;
  pimEnablement: Date;
  implementationVerified: Date;

  selected?: boolean;

}